Sequelize added

+++ server/controllers/user.controller.js
@@ -6,7 +6,8 @@ module.exports.createUser = async (request, response) => {
     const user = await User.create({ userName, email });
     response.json(user);
   } catch (err) {
-    response.status(400).json(err);
+    response.status(500).json({message: 'No se pudo crear el usuario'}); 
+    //response.status(500).json(err); 
   }
 };
 
@@ -15,7 +16,7 @@ module.exports.getAllUsers = async (_, response) => {
     const users = await User.findAll();
     response.json(users);
   } catch (err) {
-    response.json(err);
+    response.status(500).json(err);
   }
 };
 
@@ -24,28 +25,43 @@ module.exports.getUser = async (request, response) => {
     const user = await User.findOne({ where: { id: request.params.id } });
     response.json(user);
   } catch (err) {
-    response.json(err);
+    response.status(500).json(err);
   }
 };
 
 module.exports.updateUser = async (request, response) => {
   try {
-    const updatedUser = await User.update(request.body, {
-      where: { id: request.params.id },
-      returning: true,
-      plain: true
+    // Se actualiza el usuario
+    const [updatedRowCount] = await User.update(request.body, {
+      where: { id: request.params.id }
     });
-    response.json(updatedUser[1]);
+
+    // Se verifica si se ha actualizado algún registro
+    if (updatedRowCount) {
+      // Recupera la información actualizada del usuario
+      const updatedUser = await User.findOne({ where: { id: request.params.id } });
+      response.json(updatedUser);
+    } else {
+      response.status(404).json({ message: "Usuario no encontrado" });
+    }
   } catch (err) {
-    response.json(err);
+    response.status(500).json(err);
   }
 };
 
+
 module.exports.deleteUser = async (request, response) => {
   try {
+    const user = await User.findOne({ where: { id: request.params.id } });
+    if (!user) {
+      return response.status(404).json({ message: "Usuario no encontrado" });
+    }
+
     await User.destroy({ where: { id: request.params.id } });
-    response.json({ message: "User deleted" });
+
+    response.json(user);
   } catch (err) {
-    response.json(err);
+    response.status(500).json(err);
   }
 };
+
